Add files via upload
This was a challenge question given to us at Tech Elevator that I chose to work on a little bit longer so that it not only converted decimal integers to binary but also decimal fractions! Maybe I will try to create something similar that converts binary to decimal in the future.